diag; run-test-recursively;
* diag

diag is now more like a format.
either form is acceptable:

 (diag obj)
 (diag control-string &rest objs)

* run-test-recursively

Walk down through the package hierarchy under `package' and
recursively run all the tests in those packages.

e.g. it considers `:pkg.a' as a module of `:pkg', and `:pkg.a.b'
as a module of `pkg.a' and so on.

it correctly increases the indentation.