Fix finalize call in `run-test-package'
In run-test-package, finalize is called with the current package. That means
it's looking for *counter* in *package* rather than in the package running
tests. Using let, we bind *package* temporarily to the target package and
call finalize. Consequently, we also have access to *plan* from this
package, so we don't need to call in caller's *package* and pollute it.

Signed-off-by: Julien Danjou <julien@danjou.info>